requirements.txt
django>=5.0
djangorestframework>=3.15
requests>=2.32
